payroll_management/src/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
aiofiles
APScheduler
asyncpg
bcrypt
databases
email-validator
fastapi-mail
gunicorn
Jinja2
jose
passlib
pydantic
pydantic-settings
pytz
SQLAlchemy
PyJWT
python-jose
python-dateutil
python-multipart
